More model changes,moving and renaming
