--- Services/Administration/OJS.Services.Administration.Models/ContestCategories/ContestCategoriesAdministrationModel.cs
@@ -1,5 +1,0 @@
-﻿namespace OJS.Services.Administration.Models.ContestCategories;
-
-public class ContestCategoriesAdministrationModel
-{
-}
